feat:#30 scrap Api 위치 변경 + param 변경 사항 적용 setscrap,unscrap의 사용 위치변경에 따른 수정
