+++ app/src/main/java/com/dev/briefing/data/api/BriefingApi.kt
@@ -2,6 +2,12 @@ package com.dev.briefing.data.api
 
 import com.dev.briefing.data.model.BriefingDetailResponse
 import com.dev.briefing.data.model.BriefingResponse
+import com.dev.briefing.data.model.CommonResponse
+import com.dev.briefing.data.model.SetScrapRequest
+import com.dev.briefing.data.model.SetScrapResponse
+import com.dev.briefing.data.model.UnScrapResponse
+import retrofit2.http.Body
+import retrofit2.http.DELETE
 import retrofit2.http.GET
 import retrofit2.http.Path
 import retrofit2.http.Query
@@ -17,4 +23,25 @@ interface BriefingApi {
     suspend fun getBriefingDetail(
         @Path("id")id:Int,
     ): BriefingDetailResponse
+
+    /**
+     * 스크랩 등록하기
+     * @Body memberId
+     * @Body briefingId
+     */
+    @GET("/scraps/briefings")
+    suspend fun setScrap(
+        @Body memberInfo: SetScrapRequest,
+    ): CommonResponse<SetScrapResponse>
+
+    /**
+     * 스크랩 해제하기
+     * @@Path scrapId
+     */
+    @DELETE("/scraps/briefings/{briefingId}/members/{memberId}")
+    suspend fun setUnScrap(
+        @Path("briefingId") briefingId: Int,
+        @Path("memberId") memberId: Int,
+    ): CommonResponse<UnScrapResponse>
+
 }
