+++ app/src/main/java/com/dev/briefing/data/api/ScrapApi.kt
@@ -11,27 +11,6 @@ import retrofit2.http.GET
 import retrofit2.http.Path
 
 interface ScrapApi {
-
-    /**
-     * 스크랩 등록하기
-     * @Body memberId
-     * @Body briefingId
-     */
-    @GET("/scraps/briefings")
-    suspend fun setScrap(
-        @Body memberInfo: SetScrapRequest,
-    ): CommonResponse<SetScrapResponse>
-
-    /**
-     * 스크랩 해제하기
-     * @@Path scrapId
-     */
-    @DELETE("/scraps/{scrapId}")
-    suspend fun setUnScrap(
-        @Path("scrapId") scrapId: Int,
-    ): CommonResponse<UnScrapResponse>
-
-
     /**
      * 스크랩 리스트 반환하기
      * @Path memberId
