+++ app/src/main/java/com/dev/briefing/data/datasource/ScrapDataSource.kt
@@ -8,7 +8,6 @@ import com.dev.briefing.data.model.SetScrapResponse
 import com.dev.briefing.data.model.UnScrapResponse
 
 interface ScrapDataSource {
-    suspend fun setScrap(memberInfo: SetScrapRequest): CommonResponse<SetScrapResponse>
-    suspend fun setUnScrap(scrapId: Int): CommonResponse<UnScrapResponse>
+
     suspend fun getScrap(memberId: Int): CommonResponse<List<ScrapResponse>>
 }
