+++ app/src/main/java/com/dev/briefing/data/respository/ScrapRepository.kt
@@ -7,7 +7,6 @@ import com.dev.briefing.data.model.SetScrapResponse
 import com.dev.briefing.data.model.UnScrapResponse
 
 interface ScrapRepository {
-    suspend fun setScrap(memberInfo: SetScrapRequest): CommonResponse<SetScrapResponse>
-    suspend fun unScrap(scrapId:Int): CommonResponse<UnScrapResponse>
+
     suspend fun getScrap(memberId: Int): CommonResponse<List<ScrapResponse>>
 }
