+++ app/src/main/java/com/dev/briefing/data/respository/ScrapRepositoryImpl.kt
@@ -8,13 +8,7 @@ import com.dev.briefing.data.model.SetScrapResponse
 import com.dev.briefing.data.model.UnScrapResponse
 
 class ScrapRepositoryImpl(private val scrapDataSource: ScrapDataSource): ScrapRepository {
-    override suspend fun setScrap(memberInfo: SetScrapRequest): CommonResponse<SetScrapResponse> {
-        return scrapDataSource.setScrap(memberInfo)
-    }
 
-    override suspend fun unScrap(scrapId: Int): CommonResponse<UnScrapResponse> {
-        return scrapDataSource.setUnScrap(scrapId)
-    }
 
     override suspend fun getScrap(memberId: Int): CommonResponse<List<ScrapResponse>> {
         return scrapDataSource.getScrap(memberId)
