+++ app/src/main/java/com/dev/briefing/presentation/scrap/ScrapViewModel.kt
@@ -12,6 +12,8 @@ import com.dev.briefing.data.model.ScrapResponse
 import com.dev.briefing.data.model.SetScrapRequest
 import com.dev.briefing.data.respository.AuthRepository
 import com.dev.briefing.data.respository.ScrapRepository
+import com.dev.briefing.util.MEMBER_ID
+import com.dev.briefing.util.MainApplication
 import com.dev.briefing.util.SERVER_TAG
 import com.dev.briefing.util.SharedPreferenceHelper
 import kotlinx.coroutines.launch
@@ -23,6 +25,7 @@ class ScrapViewModel(private val repository: ScrapRepository) : ViewModel() {
         MutableLiveData<MutableList<ScrapResponse>>(mutableListOf())
     val scrapList: LiveData<MutableList<ScrapResponse>>
         get() = _scrapList
+    val memberId:Int = MainApplication.prefs.getSharedPreference(MEMBER_ID,0)
 
     init {
         getScrapData()
@@ -30,7 +33,7 @@ class ScrapViewModel(private val repository: ScrapRepository) : ViewModel() {
 
     /**
      * 스크랩한 기사들을 가져오는 메소드
-     * [_scraList] 업데이트
+     * [_scrapList] 업데이트
      */
     fun getScrapData() {
         viewModelScope.launch {
@@ -50,37 +53,5 @@ class ScrapViewModel(private val repository: ScrapRepository) : ViewModel() {
             }
         }
     }
-    //TODO: 스크랩한 api 결과에 따른 분기처리 혹은 return 값 수정 필요
-    fun setScrap() {
-        viewModelScope.launch {
-            try {
-                val response = repository.setScrap(
-                    memberInfo = SetScrapRequest(
-                        memberId = 0,
-                        briefingId = 0
-                    )
-                )
-                Log.d(SERVER_TAG, "통신 끝")
-                Log.d(SERVER_TAG, response.toString())
-                Log.d(SERVER_TAG, "메소드 끝")
-            } catch (e: Throwable) {
-                Log.d(SERVER_TAG, e.toString())
-            }
-        }
-    }
-    //TODO: 스크랩한 api 결과에 따른 분기처리 혹은 return 값 수정 필요
-    fun setUnScrap() {
-        viewModelScope.launch {
-            try {
-                val response = repository.unScrap(
-                    scrapId = 0
-                )
-                Log.d(SERVER_TAG, "통신 끝")
-                Log.d(SERVER_TAG, response.toString())
-                Log.d(SERVER_TAG, "메소드 끝")
-            } catch (e: Throwable) {
-                Log.d(SERVER_TAG, e.toString())
-            }
-        }
-    }
+
 }
