+++ app/src/main/java/com/dev/briefing/util/PreferenceUtil.kt
@@ -11,6 +11,7 @@ import java.time.LocalDate
 import java.time.format.DateTimeFormatter
 const val PREF_NAME = "prference"
 const val JWT_TOKEN = "token"
+const val MEMBER_ID = "memberId"
 const val SCRAP_DATE_ID = "scrap_with_date_id"
 const val SCRAP_ID_NEWS = "scrap_with_id_news"
 const val ALARM_TIME = "alarm"
